Remove use of singleton to store User 2 data
Currently the User 2 data generated in the Compare Screen is
stored in a singleton class and persists throughout the entire
lifecycle of the app. This means that if the app user signs in
and removes the User 1 associated account and then signs in
with a different User 1 account, the Compare Screen will show
the comparison for the original User 2.

Since the app regenerates the User 1 data on the User Screen
each time the app is left and returned to, it makes sense for
the User 2 data to be local only to the Compare Screen.